backend/services/ai_voice.py
import os
from dotenv import load_dotenv
from utils.helpers import make_filename
import requests

load_dotenv()

# This module provides a simple wrapper for TTS. Replace with ElevenLabs, Google, or other service.
TTS_API_KEY = os.getenv("TTS_API_KEY")

async def text_to_speech(text: str) -> str:
    """
    Convert `text` to audio and return a URL or path.
    For simplicity this example returns None or a placeholder.
    """
    # Implement according to the TTS provider you choose.
    # Example: ElevenLabs, Amazon Polly, Google TTS, etc.
    # Here we return None to indicate not available in this template.
    return None
