backend/requirements.txt
fastapi>=0.105.0
pydantic>=2.0.3
uvicorn[standard]==0.22.0
openai==1.32.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.1.0
